// apps/web/src/lib/api/generated/allauth.ts
export interface components {
  schemas: {
// ...
    AuthenticationResponse: {
      data: {
        flows: components["schemas"]["Flow"][];
      };
      meta: components["schemas"]["AuthenticationMeta"];
      /** @enum {integer} */
      status: 401;
    };
// ...
    ResetPassword: {
      /**
       * @description The password reset key
       * @example 2f-c4nqd4-e07d9bc694f9f28cd4fe92569d495333
       */
      key: string;
      password: components["schemas"]["Password"];
    };
// ...
    User: {
      /**
       * @description A neutral display value.
       * @example student@example.test
       */
      display: string;
      /**
       * @description The primary email address.
       * @example student@example.test
       */
      email: string;
      /**
       * @description Whether the account has a password.
       * @example true
       */
      has_usable_password: string;
      /**
       * @description The user UUID.
       * @example 00000000-0000-0000-0000-000000000000
       */
      id: string;
    };
    /** @description Configuration of the Django `allauth.usersessions` app. */
    UserSessionsConfiguration: {
      /** @description Matches `settings.USERSESSIONS_TRACK_ACTIVITY`. */
      track_activity: boolean;
    };
    /**
     * @description The username.
     *
     * @example wizard
     */
    Username: string;
    VerifyEmail: {
      /**
       * @description The email verification key
       * @example 2f-c4nqd4-e07d9bc694f9f28cd4fe92569d495333
       */
      key: string;
    };
// ...
  };
// ...
}